Found some small but LARGE optimizations on the code

+++ rust/src/constants.rs
@@ -20,5 +20,5 @@ pub struct Posting {
     pub n: u32,
     pub base: u32,
     pub payload: Vec<u8>,
-    pub exceptions: Option<Vec<u16>>,
+    pub exceptions: Option<Vec<u8>>,
 }
